requirements.txt
langchain 
langchain_community
ipykernel
python-dotenv
arxiv
pymupdf
pypdf
bs4
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_groq
fastapi
uvicorn
langserve
sse_starlette